+++ test/videdal/postgres_types.ex
@@ -1,0 +1,5 @@
+Postgrex.Types.define(
+  Videdal.PostgresTypes,
+  [Geo.PostGIS.Extension] ++ Ecto.Adapters.Postgres.extensions(),
+  json: Jason
+)